feat(discover) Add modal for new column editor
Add the modal window component for the new column editor. This component
takes a list of columns and fires an onApply prop when the changes are
ready to be applied.

Once this component is hooked up to the discover results table, we'll
update the EventView and navigate to the new results.